@@ -4,70 +4,72 @@
   "type": "module",
   "private": true,
   "scripts": {
-    "fetch-dev-types": "openapi-typescript https://pensjonskalkulator-backend.intern.dev.nav.no/v3/api-docs/current -o ./src/types/schema.d.ts",
-    "format:types": "prettier --write ./src/types/schema.d.ts && eslint ./src/types/schema.d.ts --fix",
-    "testc": "jest --c",
-    "test": "jest",
     "build": "next build && npm run build:server",
     "build:prod": "cp .env.prod .env && next build && npm run build:server",
     "build:server": "tsc -p tsconfig.server.json",
     "preview": "next start",
-    "dev": "concurrently \"next dev\" \"json-server mockdb/db.json --port 3001 \"",
+    "dev": "concurrently \"next dev --turbopack\" \"json-server mockdb/db.json --port 3001 \"",
+    "fetch-dev-types": "openapi-typescript https://pensjonskalkulator-backend.intern.dev.nav.no/v3/api-docs/current -o ./src/types/schema.d.ts",
+    "format:types": "prettier --write ./src/types/schema.d.ts && eslint ./src/types/schema.d.ts --fix",
+    "test": "jest",
+    "testc": "jest --c",
+    "typecheck": "tsc --noEmit --pretty -p tsconfig.json && tsc --noEmit --pretty -p tsconfig.server.json",
     "lint": "next lint",
+    "prettier:fix": "prettier --write '**/*.{cjs,js,jsx,ts,tsx,json}'",
     "prepare": "husky install"
   },
   "dependencies": {
-    "@navikt/ds-css": "^7.10.0",
-    "@navikt/ds-react": "^7.10.0",
-    "@navikt/nav-dekoratoren-moduler": "^3.1.3",
-    "@navikt/oasis": "^3.4.0",
+    "@navikt/ds-css": "^7.22.0",
+    "@navikt/ds-react": "^7.22.0",
+    "@navikt/nav-dekoratoren-moduler": "^3.2.3",
+    "@navikt/oasis": "^3.8.0",
     "@navikt/steng-for-regulering": "^0.0.7",
     "express": "^5.1.0",
-    "highcharts": "^11.4.8",
-    "highcharts-react-official": "^3.2.1",
+    "highcharts": "^12.2.0",
+    "highcharts-react-official": "^3.2.2",
     "immer": "^10.1.1",
-    "next": "14.2.22",
-    "react": "^18",
-    "react-dom": "^18",
+    "next": "15.3.3",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "server-only": "^0.0.1"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.25.4",
-    "@babel/preset-react": "^7.24.7",
-    "@babel/preset-typescript": "^7.24.7",
-    "@testing-library/jest-dom": "^6.5.0",
-    "@testing-library/react": "^16.0.1",
-    "@testing-library/user-event": "^14.5.2",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
     "@trivago/prettier-plugin-sort-imports": "^5.2.2",
     "@types/express": "^5.0.2",
-    "@types/jest": "^29.5.13",
+    "@types/jest": "^29.5.14",
     "@types/jest-axe": "^3.5.9",
-    "@types/node": "^20.17.50",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "@types/testing-library__react": "^10.2.0",
-    "@types/testing-library__react-hooks": "^4.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.8.0",
-    "@typescript-eslint/parser": "^8.8.0",
-    "@typescript-eslint/typescript-estree": "^8.8.1",
+    "@types/node": "^22.15.29",
+    "@types/react": "19.1.6",
+    "@types/react-dom": "19.1.5",
     "babel-jest": "^29.7.0",
-    "concurrently": "^9.0.1",
-    "eslint": "^8.57.1",
-    "eslint-config-next": "14.2.22",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
-    "husky": "^8.0.0",
+    "concurrently": "^9.1.2",
+    "eslint": "^9.28.0",
+    "eslint-config-next": "15.3.3",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-import-resolver-typescript": "^4.4.2",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-prettier": "^5.4.1",
+    "husky": "^9.1.7",
     "jest": "^29.7.0",
-    "jest-axe": "^9.0.0",
+    "jest-axe": "^10.0.0",
     "jest-environment-jsdom": "^29.7.0",
     "json-server": "^1.0.0-beta.3",
-    "lint-staged": "^15.2.10",
-    "openapi-typescript": "^7.4.1",
-    "prettier": "^3.3.3",
-    "typescript": "^5"
+    "lint-staged": "^16.1.0",
+    "openapi-typescript": "^7.8.0",
+    "prettier": "^3.5.3",
+    "typescript": "^5",
+    "typescript-eslint": "^8.33.1"
   },
   "lint-staged": {
-    "**/*.{cjs,js,jsx,ts,tsx}": "eslint",
+    "**/*.{cjs,js,jsx,ts,tsx}": "next lint --fix --file",
     "**/*.{cjs,js,jsx,ts,tsx,css}": "prettier --write"
+  },
+  "overrides": {
+    "@types/react": "19.1.6",
+    "@types/react-dom": "19.1.5"
   }
 }
